places_remember/requirements.txt
Django==4.1.3
psycopg2-binary==2.9.5
django-leaflet==0.28.3
social-auth-app-django==5.0.0
isort==5.10.1
django-crispy-forms==1.14.0
factory-boy==3.2.1